chore:Added review changes and facility report id
