+++ bc_obps/reporting/api/__init__.py
@@ -5,5 +5,8 @@ from .build_form_schema import build_form_schema
 from .operations import get_dashboard_operations_list
 from .activity_data import get_initial_activity_data
 from .facility_report import get_facility_report_form_data
+from .facility_report import get_facility_report_by_version_id
 from .fuel import get_fuel_data
+from .report_person_responsible import get_report_person_responsible_by_version_id
+from .report_person_responsible import save_report_contact
 from .report_activity import save_report_activity_data
