+++ bc_obps/service/report_service.py
@@ -1,14 +1,11 @@
-from typing import cast
 from uuid import UUID
 from django.db import transaction
 from registration.models import Activity, RegulatedProduct
 from registration.models.operation import Operation
-from reporting.models import ReportPersonResponsible
 from reporting.models.report import Report
 from reporting.models.facility_report import FacilityReport
 from reporting.models.report_operation import ReportOperation
 from reporting.models.report_version import ReportVersion
-from reporting.schema.report_contact import ReportPersonResponsibleIn, ReportPersonResponsibleOut
 from reporting.schema.report_operation import ReportOperationIn
 from service.data_access_service.facility_service import FacilityDataAccessService
 from service.data_access_service.report_service import ReportDataAccessService
@@ -106,33 +103,3 @@ class ReportService:
         report_operation.save()
 
         return report_operation
-
-    @classmethod
-    def get_report_person_responsible_by_version_id(cls, report_version_id: int) -> ReportPersonResponsible:
-        return ReportPersonResponsible.objects.get(report_version__id=report_version_id)
-
-    @classmethod
-    def save_report_contact(cls, version_id: int, data: ReportPersonResponsibleIn) -> ReportPersonResponsibleOut:
-        print(f"Version ID: {version_id}")
-        report_version = ReportVersion.objects.filter(id=version_id).first()
-
-        if report_version is None:
-            raise ValueError("ReportVersion with this ID does not exist.")
-
-        report_person_responsible, created = ReportPersonResponsible.objects.update_or_create(
-            report_version=report_version,
-            defaults={
-                'first_name': data.first_name,
-                'last_name': data.last_name,
-                'email': data.email,
-                'phone_number': data.phone_number,
-                'position_title': data.position_title,
-                'business_role': data.business_role,
-                'street_address': data.street_address,
-                'municipality': data.municipality,
-                'province': data.province,
-                'postal_code': data.postal_code,
-            },
-        )
-
-        return cast(ReportPersonResponsibleOut, report_person_responsible)
